Fixed tests and removed sbt-git
sbt-git is behaving in an unexpected fashion.  It is not generating
SHA-based versions; instead it always generates timestamp-based
versions.  Additionally, sbt-git was interacting badly with the scripted
sbt tests.  For now, reverted to original versioning scheme.

Also, after the PlayExceptions refactor forgot to update the associated
tests.  Fixed.

Locally all tests and scripted tests are passing.